requirements.txt
langchain
langchain-groq
langchain-community
langchain-text-splitters
langchain-classic
faiss-cpu
pypdf
huggingface-hub
sentence-transformers
flask
flask-cors
python-dotenv
gunicorn
